% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_test.R
\name{fit}
\alias{fit}
\title{frequency increment test (FIT) from Feder et al 2014}
\usage{
fit(time, v, alternative = "two.sided")
}
\arguments{
\item{time}{an integer vector of time coordinates for the time-series}

\item{v}{an integer vector of variant frequencies, i.e. counts of this variant at each time coordinate}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}
}
\value{
A two-column data frame with the FIT statistic and its p-value
}
\description{
Computes the frequency increment test (FIT) from Feder, A. F., Kryazhimskiy, S., & 
Plotkin, J. B. (2014). Identifying signatures of selection in genetic time 
series. Genetics, 196(2), 509-522. <https://doi.org/10.1534/genetics.113.158220>.  
The frequency increment test (FIT) rejects neutrality if the distribution of 
normalized variant frequency increments exhibits a mean that deviates 
significantly from zero.
}
