% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_llh.R
\name{get_llh}
\alias{get_llh}
\title{Get the log-likelihood of data while incorporating sampling error}
\usage{
get_llh(prior, lgth, tvec, nvec, bvec, prec, maxiter, verbose, s, alpha, f0)
}
\arguments{
\item{prior}{a function that returns the prior distribution (default: function(x){1})}

\item{lgth}{number of data points in the time series}

\item{tvec}{time points for the time series (start with 0)}

\item{nvec}{total number of samples for each time point}

\item{bvec}{number of successful samples at each time point}

\item{prec}{minimum precision for optimisation}

\item{maxiter}{maximum number of iterations for optimisation}

\item{verbose}{print intermediate output (TRUE or FALSE)}

\item{s}{selection coefficient}

\item{alpha}{population size, bounded [0,Inf)}

\item{f0}{initial frequency for the underlying logistic change}
}
\value{
A log-likelihood estimate.
}
\description{
\code{get_llh} uses Laplace approximation to compute the log-likelihood of the gaussian approximation assuming the data is drawn from a series of binomial choices.
}

