% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_llh_freq.R
\name{get_llh_freq}
\alias{get_llh_freq}
\title{Get the log-likelihood of data assuming sample frequencies are actual frequencies}
\usage{
get_llh_freq(prior, lgth, prec, tvec, nuvec, s, alpha, f0)
}
\arguments{
\item{prior}{a function that returns the prior distribution (default: function(x){1})}

\item{lgth}{number of data points in the time series}

\item{prec}{minimum precision for optimisation}

\item{tvec}{time points for the time series (start with 0)}

\item{nuvec}{percent of successful samples per time point}

\item{s}{selection coefficient}

\item{alpha}{population size, bounded [0,Inf)}

\item{f0}{initial frequency for the underlying logistic change}
}
\value{
A log-likelihood estimate.
}
\description{
\code{get_llh_freq} computes the Gaussian approximation assuming that the actual frequencies are the sampled frequencies
}

