% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsinfer.R
\name{tsinfer}
\alias{tsinfer}
\title{Infer selection/drift parameters for time-series allele-frequency data}
\usage{
tsinfer(tvec, bvec, nvec, maxiter = 200, prec = 1e-06, iffreq = FALSE,
  ifneut = FALSE, iffixedf0 = FALSE, verbose = FALSE)
}
\arguments{
\item{tvec}{Time coordinates for the time-series (start with 0)}

\item{bvec}{Number of new form per time point}

\item{nvec}{Total number of samples per time point}

\item{maxiter}{Maximum number of iterations}

\item{prec}{Precision for optimisation}

\item{iffreq}{whether to assume that the sample frequencies are the population frequncies (default=FALSE)}

\item{ifneut}{whether to compute only the neutral model (default=FALSE)}

\item{iffixedf0}{whether to use the initial sample frequency as the initial frequency of the logistic component of the model (default=FALSE)}

\item{verbose}{whether to print intermediate output (detault=FALSE)}
}
\value{
A list with the neutral and non-neutral parameter values and associated log-likelihoods
}
\description{
\code{tsinfer} computes the selection coefficient, population size and initial frequency for the logistic using a Gaussian approximation of the Kimura expression for time-series allele-frequency data.
}

