% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsinfer.R
\name{tsinfer}
\alias{tsinfer}
\title{Infer selection/drift parameters for time-series allele-frequency data}
\usage{
tsinfer(tvec, bvec, nvec, maxiter = 200, prec = 1e-06, iffreq = FALSE,
  ifneut = FALSE, iffixedf0 = FALSE, verbose = FALSE, mins = -2,
  maxs = 2, minalpha = 10, maxalpha = 1e+08, minf0 = -10, maxf0 = 10)
}
\arguments{
\item{tvec}{Time coordinates for the time-series (start with 0)}

\item{bvec}{Number of new form per time point}

\item{nvec}{Total number of samples per time point}

\item{maxiter}{Maximum number of iterations}

\item{prec}{Precision for optimisation}

\item{iffreq}{whether to assume that the sample frequencies are the population frequencies (default=FALSE)}

\item{ifneut}{whether to compute only the neutral model (default=FALSE)}

\item{iffixedf0}{whether to use the initial sample frequency as the initial frequency of the logistic component of the model (default=FALSE)}

\item{verbose}{whether to print intermediate output (detault=FALSE)}

\item{mins}{minimum s value to consider (default=-2)}

\item{minalpha}{minimum alpha value to consider (default=10)}

\item{minf0}{minimum f0 value to consider in log-odds (default=-10)}

\item{maxf0}{maximum f0 value to consider in log-odds (default=10)}

\item{mins}{maximum s value to consider (default=2)}

\item{minalpha}{maximum alpha value to consider (default=1e8)}
}
\value{
A list with the neutral and non-neutral parameter values and associated log-likelihoods
The output of the execution is the following list:
s = selection coefficient for non-neutral model
alpha = population size for non-neutral model
f0 = initial frequency for the logistic in non-neutral model
LL = log-likelihood of non-neutral model
s.0 = 0 (selection coefficient for the neutral model)
alpha.0 = population size for the neutral model
f0.0 = initial frequency for the logistic in neutral model
LL.0 = log-likelihood of neutral model
}
\description{
\code{tsinfer} computes the selection coefficient, population size and initial frequency for the logistic using a Gaussian approximation of the Kimura expression for time-series allele-frequency data.
}
